datapunk/datapunk-cortex/pyproject.toml
[tool.poetry]
name = "datapunk-cortex"
version = "0.1.0"
description = "Datapunk Cortex Service"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
# Testing
pytest = "8.3.3"
pytest-asyncio = "0.24.0"
pytest-cov = "4.1.0"
httpx = "0.27.0"

# ML/DL
torch = "^2.2.0"
transformers = "4.37.2"
numpy = "^1.24.0"
scikit-learn = "^1.3.0"

# NLP & Search
farm-haystack = "^1.21.0"
langchain = "^0.1.0"
spacy = "^3.7.2"

# API
fastapi = "0.109.1"
uvicorn = "0.27.0"
pydantic = "2.6.1"
pydantic-settings = "2.1.0"

# Database & Storage
redis = "5.0.1"
sqlalchemy = "2.0.25"
psycopg2-binary = "2.9.9"
minio = "7.2.3"
pyarrow = "^14.0.1"

# Utilities
python-dotenv = "0.0.0"
python-multipart = "0.0.9"
pyyaml = "6.0.1"
requests = "2.31.0"

# Monitoring
prometheus-client = "0.19.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
